+++ voicenavigator-20180612/ChangeLog.md
@@ -1,3 +1,12 @@
+2026-09-03 Version: 2.0.5
+- Update API AuditTTSVoice: add request parameters NlsServiceType.
+- Update API DescribeTTSConfig: add response parameters Body.TtsOverridesUuid.
+- Update API GetAsrConfig: add response parameters Body.Data.AsrOverridesUuid.
+- Update API ModifyAsrConfig: add request parameters AsrOverridesUuid.
+- Update API ModifyAsrConfig: add request parameters EngineXunfei.
+- Update API ModifyTTSConfig: add request parameters TtsOverridesUuid.
+
+
 2026-08-13 Version: 2.0.4
 - Update API DescribeTTSConfig: add response parameters Body.BackgroundMusicName.
 - Update API ModifyTTSConfig: add request parameters BackgroundMusicName.